labeling_ui/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
requests>=2.32.3
Pillow>=10.3.0
PyYAML>=6.0.2
